new package name

+++ src/autonomous_rover/src/feedback_node.cpp
@@ -14,7 +14,6 @@ using namespace std::chrono_literals;
 class FeedbackCalibration : public rclcpp::Node {
 public:
   FeedbackCalibration() : Node("feedback_node") {
-    // Initialize 
     wheel_radius_ = 0.0325;
     positions_.resize(4, 0.0f);
     velocities_.resize(4, 0.0f);
@@ -26,14 +25,12 @@ public:
       {"back_right_wheel_joint", 3},
     };
 
-    // Subscribers
     encoder_sub_ = this->create_subscription<sensor_msgs::msg::JointState>(
       "/joint_states", 10, std::bind(&FeedbackCalibration::encoder_callback, this, std::placeholders::_1));
     
     odom_sub_ = this->create_subscription<nav_msgs::msg::Odometry>(
       "/odometry/filtered", 10, std::bind(&FeedbackCalibration::odom_callbak, this, std::placeholders::_1));
 
-    // Feedback Publishers
     publish_timer_ = this->create_wall_timer(100ms, std::bind(&FeedbackCalibration::message_publisher, this));
   }
 
@@ -43,14 +40,11 @@ private:
 
     for (size_t i=0; i < msg->name.size();++i){
 
-      // Find the index of the encoder from the map
       auto it = umap_.find(msg->name[i]);
 
       if (it != umap_.end()){
-        // Append the encoder readings to the according joint
         size_t encoder_index = it->second;
 
-        // Convert message content to meters
         positions_[encoder_index] = msg->position[i] * wheel_radius_;
         velocities_[encoder_index] = msg->velocity[i] * wheel_radius_;
 
@@ -63,7 +57,6 @@ private:
 
   void odom_callbak(const nav_msgs::msg::Odometry::SharedPtr msg) {
 
-    // Extracting robot pose (Position and Orientation)
     robot_x_.push_back(msg->pose.pose.position.x);
     robot_y_.push_back(msg->pose.pose.position.y);
     robot_z_.push_back(msg->pose.pose.position.z);
@@ -71,12 +64,9 @@ private:
     Eigen::Quaternionf q(msg->pose.pose.orientation.w, msg->pose.pose.orientation.x, msg->pose.pose.orientation.y, 
       msg->pose.pose.orientation.z); // ( w, x, y, z )
     
-    // In format Roll Pitch Yaw
     Eigen::Vector3f rpy = q.toRotationMatrix().eulerAngles(0, 1, 2);
-    // Extract Yaw
     robot_yaw_.push_back(rpy[2] * 180.0 / M_PI);
 
-    // Extracting robot velocity
     linear_vel_.push_back(msg->twist.twist.linear.x);
     angular_vel_.push_back(msg->twist.twist.angular.z);
   }
@@ -97,27 +87,25 @@ private:
     RCLCPP_INFO(this->get_logger(), "---------------------------------");
     RCLCPP_INFO(this->get_logger(), "Encoder Error (L-R):");
     RCLCPP_INFO(this->get_logger(), "Position: %f m",((positions_[0] + positions_[1]) / 2) - ((positions_[2] + positions_[3])/2));
-    RCLCPP_INFO(this->get_logger(), "Velocities: %f m/s",((velocities_[0] + velocities_[1]) / 2) -((velocities_[2] + velocities_[3])/2));
+    RCLCPP_INFO(this->get_logger(), "Velocities: %f m/s",((velocities_[0] + velocities_[1]) / 2) - ((velocities_[2] + velocities_[3])/2));
 
     RCLCPP_INFO(this->get_logger(), "---------------------------------");
     RCLCPP_INFO(this->get_logger(), "------EKF Odometry Feedback------");
     RCLCPP_INFO(this->get_logger(), "Robot Position:");
-    RCLCPP_INFO(this->get_logger(), "X: %f m ", robot_x_.back());
-    RCLCPP_INFO(this->get_logger(), "Y: %f m/s", robot_y_.back());
-    RCLCPP_INFO(this->get_logger(), "Yaw: %f ", robot_yaw_.back());
+    RCLCPP_INFO(this->get_logger(), "X: %d m ", robot_x_.back());
+    RCLCPP_INFO(this->get_logger(), "Y: %d m/s", robot_y_.back());
+    RCLCPP_INFO(this->get_logger(), "Yaw: %d ", robot_yaw_.back());
 
     RCLCPP_INFO(this->get_logger(), "---------------------------------");
     RCLCPP_INFO(this->get_logger(), "Robot Velocity:");
-    RCLCPP_INFO(this->get_logger(), "Linear X: %f m ", linear_vel_.back());
-    RCLCPP_INFO(this->get_logger(), "Angular Z: %f m/s", angular_vel_.back());
+    RCLCPP_INFO(this->get_logger(), "Linear X: %d m ", linear_vel_.back());
+    RCLCPP_INFO(this->get_logger(), "Angular Z: %d m/s", angular_vel_.back());
 
   }
-  // Subscribers
   rclcpp::Subscription<sensor_msgs::msg::JointState>::SharedPtr encoder_sub_;
   rclcpp::Subscription<nav_msgs::msg::Odometry>::SharedPtr odom_sub_;
   rclcpp::TimerBase::SharedPtr publish_timer_;
 
-  // State variables
   std::unordered_map<std::string, size_t> umap_;
   std::vector<float> positions_;
   std::vector<float> velocities_;
@@ -126,10 +114,10 @@ private:
   std::vector<double> robot_y_;
   std::vector<double> robot_z_;
   std::vector<double> robot_yaw_;
-  std::vector<float> linear_vel_;
-  std::vector<float> angular_vel_;
+  std::vector<double> linear_vel_;
+  std::vector<double> angular_vel_;
 
-  float wheel_radius_;
+  double wheel_radius_;
 };
 
 int main(int argc, char * argv[]) {
